update lifting state lab
