+++ src/Components/Event.js
@@ -1,7 +1,15 @@
-export default function Event({ events }) {
+import { useState } from "react";
+import Attendees from "./Attendees";
+
+export default function Event({ event, updateEventAttendance, attendees }) {
+  const [showAttendees, setShowAttendees] = useState(false);
+
+  function toggleEventAttendees() {
+    setShowAttendees(!showAttendees);
+  }
+
   return (
-    <>
-      {/* Event */}
+    <div className="event">
       <li key={event.id}>
         <img src={event.eventImage} alt={event.name} />
         <h5>
@@ -11,47 +19,19 @@ export default function Event({ events }) {
         <span>Organized by: {event.organizer} </span>
         <br />
         <>
-          {/* Attendees */}
           <button onClick={toggleEventAttendees}>
             {!showAttendees ? "Show Attendees" : "Hide Attendees"}
           </button>
 
           {showAttendees ? (
-            <div className="attendees">
-              {attendees.map((attendee, index) => (
-                <>
-                  {/* Attendee */}
-                  <div key={attendee.id} className="attendee">
-                    <p>
-                      <img src={attendee.avatar} alt={attendee.firstName} />
-                      {"   "}
-                      <span>
-                        {" "}
-                        {attendee.firstName} {attendee.lastName}{" "}
-                      </span>
-                    </p>
-                    <p>
-                      <button
-                        className="clickable"
-                        onClick={() =>
-                          updateEventAttendance(event.id, attendee.id)
-                        }
-                      >
-                        Attending:
-                      </button>
-                      <span>{attendee.attendance ? "✅" : "❌"}</span>
-                    </p>
-
-                    <p>
-                      <span>Note:</span> {attendee.note}
-                    </p>
-                  </div>
-                </>
-              ))}
-            </div>
+            <Attendees
+              attendees={attendees}
+              updateEventAttendance={updateEventAttendance}
+              event={event}
+            />
           ) : null}
         </>
       </li>
-    </>
+    </div>
   );
 }
